refactor: 이미지 업로드 로직 커스텀 훅으로 추출, 필요없는 파일 삭제
