+++ features/post/create/CreatePostPage.tsx
@@ -12,7 +12,6 @@ import Modal from "@/components/Modal";
 import { useEffect, useState } from "react";
 import BgColorSelector from "./BgColorSelector";
 import Writing from "./Writing";
-// import BgImageSelector from "./BgImageSelector";
 import ImageUploader from "./ImageUploader";
 import TitleEditor from "./TitleEditor";
 import ModalWithoutCloseBtn from "@/components/ModalWithoutCloseBtn";
@@ -76,7 +75,7 @@ export default function CreatePostPage({
       break;
     case 4:
       title = "사진 추가";
-      content = <ImageUploader isMobile={false} />;
+      content = <ImageUploader />;
       break;
     case 5:
       title = "표지 미리보기";
